+++ axiom/test/test_upgrading.py
@@ -24,7 +24,7 @@ from axiom.scripts import axiomatic
 from axiom.store import Store
 from axiom.substore import SubStore
 from axiom.plugins.axiom_plugins import Upgrade
-from axiom.test.util import CommandStub
+from axiom.test.util import CommandStub, callWithStdoutRedirect
 import six
 
 
@@ -150,28 +150,6 @@ def _logMessagesFrom(f):
         log.removeObserver(L.append)
         return ign
     return d.addBoth(x).addCallback(lambda ign: L)
-
-
-
-def callWithStdoutRedirect(f, *a, **kw):
-    """
-    Redirect stdout and invoke C{f}.
-
-    @returns: C{(returnValue, stdout})
-    """
-    if six.PY3:
-        output = io.StringIO()
-    else:
-        # We need to emulate stdio implicit encoding goofiness on python 2,
-        # because some things write text to stdout and some things write bytes
-        import StringIO
-        output = StringIO.StringIO()
-    sys.stdout, stdout = output, sys.stdout
-    try:
-        result = f(*a, **kw)
-    finally:
-        sys.stdout = stdout
-    return result, output
